back-end/requirements.txt
backcall==0.1.0
elasticsearch==7.7.0
flask==1.1.2
flask-babel==1.0.0
flask-cors==3.0.8
flask-httpauth==4.0.0
flask_mail==0.9.1
flask-migrate==2.5.3
flask-socketio==4.2.1
flask-sqlalchemy==2.4.1
gevent==20.5.0
gunicorn==20.0.4
httpie==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pyppeteer==0.0.25
python-dotenv==0.13.0
PyJWT==1.7.1
redis==3.5.2
requests==2.22.0
rq==1.4.1
rq-dashboard==0.6.1
traitlets==4.3.3
wcwidth==0.1.9
